I have no idea if this is what is required.
Please tell me if I'm doing something completely wrong. The heading
given is a little cryptic so hopefully I have the right idea.

Also, "Desc of methods of evalutation of the work" seems to be the same
thing as this, I'm unsure how it should differ.

I won't bother with an excuse on why this is late but instead offer an
assurance that I will try my very best to stick to the schedule in the
future.