Dared to make some SQL; Implemented first method
getPData seemed to be missing from the SQL statements so I dared to
implement it myself by looking at the other queries for reference.
Should be functional and do what we need it to.

The first method that I've done in Database is the implementation of the
SQL I created, using Luke's template.