Re-DL'd screenshots from GitHub site
For unknown reason my Windows client was reading all of the images that
had the time with Colons as deleted.  So I've downloaded the raw form of
each on the offending images and the time uses hyphens instead.  If you
have used any of the images I'll go through the manual files and change
them for you now