Fixed Screenshots
The problem was that Louis was putting colons into the file names for
the screenshots before and windows 7 doesn't like that, so they weren't
appearing in my directory because windows wasn't accepting them.  I've
replaced the colons with hyphens and have actually checked I can open
the things this time.