Added package to manual base file; System WIP
the 'grffile' package seems to remove the text link found in the pdf
stating the image filename, which is nice.  Graphicspath doesn't seem to
work so I removed it.  Finished the general section of the Using the
System file and am proceeding to do a section for each button and
whatnot.